+++ sheraf/__init__.py
@@ -12,6 +12,7 @@ from .attributes.collections import (
     LargeListAttribute,
     ListAttribute,
     SetAttribute,
+    SmallDictAttribute,
     SmallListAttribute,
 )
 from .attributes.counter import CounterAttribute
@@ -54,7 +55,6 @@ from .exceptions import (
     EmptyQuerySetUnpackException,
     PrimaryKeyException,
 )
-from .indexes import Index
 from .models import (
     AutoModel,
     AttributeModel,